- Tiny datatype optimization
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1747 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/NPCs.cpp
@@ -255,7 +255,7 @@ void NPC::sendGetNumber(int32_t def, int32_t min, int32_t max) {
 void NPC::sendStyle(int32_t styles[], int8_t size) {
 	PacketCreator packet = npcPacket(NPCDialogs::style);
 	packet.addByte(size);
-	for (int32_t i = 0; i < size; i++)
+	for (int8_t i = 0; i < size; i++)
 		packet.addInt(styles[i]);
 	player->getSession()->send(packet);
 }
